Jarvis-Visual-Engine/requirements.txt
# Core
python-dotenv==1.0.0
PyYAML==6.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Camera & Vision
opencv-python==4.8.1.78
face-recognition==1.3.5
Pillow==10.1.0
numpy==1.24.3

# Vision APIs
openai==1.3.7
anthropic==0.7.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
timescaledb==0.0.15
alembic==1.13.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# Caching
redis==5.0.1

# Web & API
flask==3.0.0
flask-cors==4.0.0
flask-restx==0.5.1
python-socketio==5.10.0
python-engineio==4.8.0
python-multipart==0.0.6

# ONVIF Camera Support
zeep==4.2.1
onvif-zeep==0.2.12
onvif==0.2.12

# Multi-object tracking
torch==2.1.1
torchvision==0.16.1

# Utilities
requests==2.31.0
cryptography==41.0.7
python-dateutil==2.8.2
Werkzeug==3.0.1
tenacity==8.2.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
